Fix home ownernship defaulting to root
The mfg tool unpacks all files of the rootfs with root umask.
This leads to wrong owenrships for the /home directories.

The fix implemented consists of a service that tries to restore the ownerships inside
the /home directories. This is very likley only an bandaid fix for the problem.

It is likley that other issues will appear in the future that are caused by
the mfg flash process unpacking method. A direct solution implemented inside the
flash process may fix the problem completly.